requirements.txt
celery==5.0.2
django==3.1.3
django-celery-results==2.0.0
django-celery-beat==2.1.0
djangorestframework==3.12.1
djangorestframework-filters==1.0.0.dev2
django-cors-headers==3.5.0
psycopg2-binary==2.8.6
pycurl==7.43.0.6
python-decouple==3.3
pytz==2020.4
redis==3.5.3
requests==2.24.0
gunicorn==20.0.4
